Add files via upload
Create a HTMl page with a paragraph written on it and under which 9 buttons are placed in a 3x3 grid. The first row is for buttons labled with colors names Red,Green and Blue,second row with numbers 10,20,30, and the third row with different font names. Click event of each of the buttons should make appropriate changes in the style of the paragraph